requirements.txt
asgiref==3.6.0
mysqlclient
Django==3.2.18
djangorestframework==3.14.0
Pillow==9.5.0
PyMySQL==1.0.3
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.5.0